requirements.txt
quart
langchain
langgraph
python-dotenv
fastmcp
ollama
ollama
